# Don't return negative zeros from long decimal operations