--- a/core/trino-main/src/test/java/io/trino/operator/scalar/AbstractTestFunctions.java
+++ b/core/trino-main/src/test/java/io/trino/operator/scalar/AbstractTestFunctions.java
@@ -99,7 +99,7 @@ public abstract class AbstractTestFunctions
         functionAssertions.assertFunction(format("\"%s\"(%s)", mangleOperatorName(operator), value), expectedType, expected);
     }
 
-    protected void assertDecimalFunction(String statement, SqlDecimal expectedResult)
+    protected void assertDecimalFunction(@Language("SQL") String statement, SqlDecimal expectedResult)
     {
         assertFunction(
                 statement,
@@ -112,12 +112,12 @@ public abstract class AbstractTestFunctions
         functionAssertions.assertInvalidFunction(projection, errorCode, message);
     }
 
-    protected void assertInvalidFunction(String projection, String message)
+    protected void assertInvalidFunction(@Language("SQL") String projection, String message)
     {
         functionAssertions.assertInvalidFunction(projection, INVALID_FUNCTION_ARGUMENT, message);
     }
 
-    protected void assertInvalidFunction(String projection, ErrorCodeSupplier expectedErrorCode)
+    protected void assertInvalidFunction(@Language("SQL") String projection, ErrorCodeSupplier expectedErrorCode)
     {
         functionAssertions.assertInvalidFunction(projection, expectedErrorCode);
     }