--- a/core/trino-spi/src/main/java/io/trino/spi/type/LongDecimalType.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/type/LongDecimalType.java
@@ -26,6 +26,8 @@ import io.trino.spi.function.BlockIndex;
 import io.trino.spi.function.BlockPosition;
 import io.trino.spi.function.ScalarOperator;
 
+import java.math.BigInteger;
+
 import static io.airlift.slice.SizeOf.SIZE_OF_LONG;
 import static io.trino.spi.block.Int128ArrayBlock.INT128_BYTES;
 import static io.trino.spi.function.OperatorType.COMPARISON;
@@ -35,6 +37,7 @@ import static io.trino.spi.type.Decimals.decodeUnscaledValue;
 import static io.trino.spi.type.TypeOperatorDeclaration.extractOperatorDeclaration;
 import static io.trino.spi.type.UnscaledDecimal128Arithmetic.UNSCALED_DECIMAL_128_SLICE_LENGTH;
 import static io.trino.spi.type.UnscaledDecimal128Arithmetic.compare;
+import static io.trino.spi.type.UnscaledDecimal128Arithmetic.isNegative;
 import static java.lang.invoke.MethodHandles.lookup;
 
 final class LongDecimalType
@@ -95,7 +98,10 @@ final class LongDecimalType
             return null;
         }
         Slice slice = getSlice(block, position);
-        return new SqlDecimal(decodeUnscaledValue(slice), getPrecision(), getScale());
+        BigInteger unscaledValue = decodeUnscaledValue(slice);
+        // sanity check: disallow "negative zero" in case we get a corrupted result from somewhere
+        checkArgument(unscaledValue.signum() != 0 || !isNegative(slice), "The encoded long decimal is invalid: negative zero");
+        return new SqlDecimal(unscaledValue, getPrecision(), getScale());
     }
 
     @Override