When doing multiplication, division and modulus of long decimals, the negative flag on the result was set unconditionally whenever the result was expected to be negative based on the signedness of the operands, but it didn't take zeros into consideration. So in certain situations the result was a negative zero, e.g. when multiplying zero by a negative number. We fix this by making the `pack` method "correct" the negative bit when the number to pack is zero, which is the very last stage of each of these operations.

One particular case this caused problems was when the optimizer was evaluating a constant expression, because the `Slice` it was encoded into was not unpacked (unpacking would neutralize the sing bit) and instead used for further evaluation. So the following query was wrongly returning `false` because the negative zero was compared to regular zero and the test was done bit-by-bit:

```sql
SELECT CAST(0 AS decimal(30,5)) = CAST(0 AS decimal(30,5)) * -1;
```

We also add an explicit check for negative zero when decoding `SqlDecimal` from `Slice`. This is similar to the checks done for timestamps, and in most cases can detect silent corruption like the negative zero produced by multiplication etc.. This check allows us to detect bugs like this, because the testing infrastructure (used by `TestDecimalOperators`) implicitly converts the actual and expected values to `SqlDecimal`, which loses the sign bit when decoding from the `Slice`, so we can't detect negative zeros at this stage.

There are also some additional test cases for corner cases around negative zero.

Benchmark results (done on a representative case) to measure the performance impact of this change suggest that this is within the noise (the results are pretty stable, but the modified code actually comes up slightly faster, which I consider a fluke):

Original version:

```
Benchmark                                    (expression)   Mode   Cnt  Score   Error   Units
BenchmarkDecimalOperators.multiplyBenchmark       l1 * l2  thrpt  1000  0,266 ± 0,004  ops/ms
```

Modified version:

```
Benchmark                                    (expression)   Mode   Cnt  Score   Error   Units
BenchmarkDecimalOperators.multiplyBenchmark       l1 * l2  thrpt  1000  0,260 ± 0,004  ops/ms
```

Modified version with `||` instead of `|` in the checks for zero:

```
Benchmark                                    (expression)   Mode   Cnt  Score   Error   Units
BenchmarkDecimalOperators.multiplyBenchmark       l1 * l2  thrpt  1000  0,263 ± 0,004  ops/ms
```