@ksobolew 2021-09-01
Regarding performance implications of this change, I did run some benchmarks (as in `BenchmarkDecimalOperators`) and the difference is imperceptible and thoroughly drowned out by the noise. You can run your own benchmarks, of course, and see if you can measure it somehow. But I don't think the impact is significant.