@sopel39 2021-09-01
>  I did run some benchmarks (as in BenchmarkDecimalOperators) and the difference is imperceptible and thoroughly drowned out by the noise.

Do you have the numbers? Could you share them here?