@findepi 2021-09-01
Numbers aside, we should have correctness first. 

1. let's fix (forbid) negative zeros from the multiplcation
2. let's *separately* (i.e. as a separate commit, can be same PR), add validation on values while returning them from `DecimalType.getObjectValue`
   - see discussion https://github.com/trinodb/trino/pull/8985#discussion_r696699567 for details
   - @ksobolew validation in `UnscaledDecimal128Arithmetic#unscaledDecimalToBigInteger` is somewhat implicit, i would rather have it explicit
   - also, we don't intend to add this validation while operating on long decimal values and `UnscaledDecimal128Arithmetic#unscaledDecimalToBigInteger ` is used in more places than just `getObjectValue`.
3. we *may* follow up with relaxing type representation to allow "negative zero" as an alternative encoding for zero value (*not* a separate value). 
    - this would require changes in operators: equality, comparison, IS DISTINCT, etc.
    - first and foremost, such representation change requires a separate discussion, so let's keep it off this PR