@ksobolew 2021-09-01
> Do you have the numbers? Could you share them here?

Sure!

Before:

```
Benchmark                                             (expression)   Mode  Cnt  Score   Error   Units
BenchmarkDecimalOperators.additionBenchmark                d1 + d2  thrpt   30  6,171 ± 0,060  ops/ms
BenchmarkDecimalOperators.additionBenchmark      d1 + d2 + d3 + d4  thrpt   30  3,535 ± 0,032  ops/ms
BenchmarkDecimalOperators.additionBenchmark                s1 + s2  thrpt   30  6,418 ± 0,051  ops/ms
BenchmarkDecimalOperators.additionBenchmark      s1 + s2 + s3 + s4  thrpt   30  3,570 ± 0,040  ops/ms
BenchmarkDecimalOperators.additionBenchmark                l1 + l2  thrpt   30  0,711 ± 0,031  ops/ms
BenchmarkDecimalOperators.additionBenchmark      l1 + l2 + l3 + l4  thrpt   30  0,314 ± 0,003  ops/ms
BenchmarkDecimalOperators.additionBenchmark      s2 + l3 + l1 + s4  thrpt   30  0,356 ± 0,005  ops/ms
BenchmarkDecimalOperators.additionBenchmark              lz1 + lz2  thrpt   30  0,778 ± 0,013  ops/ms
BenchmarkDecimalOperators.additionBenchmark  lz1 + lz2 + lz3 + lz4  thrpt   30  0,320 ± 0,037  ops/ms
BenchmarkDecimalOperators.additionBenchmark    s2 + lz3 + lz1 + s4  thrpt   30  0,319 ± 0,004  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                d1 / d2  thrpt   30  5,957 ± 0,046  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      d1 / d2 / d3 / d4  thrpt   30  2,923 ± 0,034  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                i1 / i2  thrpt   30  2,962 ± 0,028  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      i1 / i2 / i3 / i4  thrpt   30  1,101 ± 0,014  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s1 / s2  thrpt   30  1,775 ± 0,172  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      s1 / s2 / s2 / s2  thrpt   30  0,824 ± 0,018  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s1 / s3  thrpt   30  0,346 ± 0,004  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s2 / l1  thrpt   30  0,574 ± 0,009  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                l1 / s2  thrpt   30  0,355 ± 0,010  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s3 / l1  thrpt   30  0,298 ± 0,003  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                l2 / l3  thrpt   30  0,221 ± 0,003  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      l2 / l4 / l4 / l4  thrpt   30  0,131 ± 0,011  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      l2 / s4 / s4 / s4  thrpt   30  0,130 ± 0,002  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  d1 % d2  thrpt   30  2,124 ± 0,017  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        d1 % d2 % d3 % d4  thrpt   30  1,216 ± 0,011  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  i1 % i2  thrpt   30  2,996 ± 0,027  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        i1 % i2 % i3 % i4  thrpt   30  1,095 ± 0,014  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  s1 % s2  thrpt   30  0,664 ± 0,012  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        s1 % s2 % s2 % s2  thrpt   30  0,220 ± 0,004  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  s2 % l2  thrpt   30  0,742 ± 0,012  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  l3 % s3  thrpt   30  0,360 ± 0,005  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  s4 % l3  thrpt   30  0,414 ± 0,007  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  l2 % l3  thrpt   30  0,306 ± 0,003  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        l2 % l3 % l4 % l1  thrpt   30  0,131 ± 0,004  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                d1 * d2  thrpt   30  6,219 ± 0,057  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark      d1 * d2 * d3 * d4  thrpt   30  3,550 ± 0,034  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                i1 * i2  thrpt   30  6,382 ± 0,042  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                s1 * s2  thrpt   30  6,418 ± 0,067  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark      s1 * s2 * s5 * s6  thrpt   30  3,379 ± 0,043  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                s3 * s4  thrpt   30  1,715 ± 0,035  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                l2 * s2  thrpt   30  1,130 ± 0,027  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark      l2 * s2 * s5 * s6  thrpt   30  0,513 ± 0,006  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                s1 * l2  thrpt   30  1,128 ± 0,019  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                l1 * l2  thrpt   30  1,025 ± 0,020  ops/ms
```

After:

```
Benchmark                                             (expression)   Mode  Cnt  Score   Error   Units
BenchmarkDecimalOperators.additionBenchmark                d1 + d2  thrpt   30  6,179 ± 0,059  ops/ms
BenchmarkDecimalOperators.additionBenchmark      d1 + d2 + d3 + d4  thrpt   30  3,531 ± 0,032  ops/ms
BenchmarkDecimalOperators.additionBenchmark                s1 + s2  thrpt   30  6,400 ± 0,051  ops/ms
BenchmarkDecimalOperators.additionBenchmark      s1 + s2 + s3 + s4  thrpt   30  3,565 ± 0,040  ops/ms
BenchmarkDecimalOperators.additionBenchmark                l1 + l2  thrpt   30  0,729 ± 0,013  ops/ms
BenchmarkDecimalOperators.additionBenchmark      l1 + l2 + l3 + l4  thrpt   30  0,311 ± 0,004  ops/ms
BenchmarkDecimalOperators.additionBenchmark      s2 + l3 + l1 + s4  thrpt   30  0,347 ± 0,005  ops/ms
BenchmarkDecimalOperators.additionBenchmark              lz1 + lz2  thrpt   30  0,766 ± 0,029  ops/ms
BenchmarkDecimalOperators.additionBenchmark  lz1 + lz2 + lz3 + lz4  thrpt   30  0,338 ± 0,003  ops/ms
BenchmarkDecimalOperators.additionBenchmark    s2 + lz3 + lz1 + s4  thrpt   30  0,312 ± 0,004  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                d1 / d2  thrpt   30  5,956 ± 0,055  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      d1 / d2 / d3 / d4  thrpt   30  2,935 ± 0,020  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                i1 / i2  thrpt   30  2,954 ± 0,023  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      i1 / i2 / i3 / i4  thrpt   30  1,105 ± 0,013  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s1 / s2  thrpt   30  1,958 ± 0,174  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      s1 / s2 / s2 / s2  thrpt   30  0,822 ± 0,016  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s1 / s3  thrpt   30  0,350 ± 0,002  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s2 / l1  thrpt   30  0,591 ± 0,013  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                l1 / s2  thrpt   30  0,354 ± 0,007  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                s3 / l1  thrpt   30  0,301 ± 0,003  ops/ms
BenchmarkDecimalOperators.divisionBenchmark                l2 / l3  thrpt   30  0,222 ± 0,003  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      l2 / l4 / l4 / l4  thrpt   30  0,140 ± 0,011  ops/ms
BenchmarkDecimalOperators.divisionBenchmark      l2 / s4 / s4 / s4  thrpt   30  0,131 ± 0,001  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  d1 % d2  thrpt   30  2,122 ± 0,018  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        d1 % d2 % d3 % d4  thrpt   30  1,213 ± 0,008  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  i1 % i2  thrpt   30  3,013 ± 0,017  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        i1 % i2 % i3 % i4  thrpt   30  1,090 ± 0,014  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  s1 % s2  thrpt   30  0,653 ± 0,015  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        s1 % s2 % s2 % s2  thrpt   30  0,227 ± 0,003  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  s2 % l2  thrpt   30  0,717 ± 0,015  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  l3 % s3  thrpt   30  0,362 ± 0,006  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  s4 % l3  thrpt   30  0,419 ± 0,006  ops/ms
BenchmarkDecimalOperators.moduloBenchmark                  l2 % l3  thrpt   30  0,318 ± 0,010  ops/ms
BenchmarkDecimalOperators.moduloBenchmark        l2 % l3 % l4 % l1  thrpt   30  0,131 ± 0,004  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                d1 * d2  thrpt   30  6,209 ± 0,053  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark      d1 * d2 * d3 * d4  thrpt   30  3,560 ± 0,034  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                i1 * i2  thrpt   30  6,337 ± 0,054  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                s1 * s2  thrpt   30  6,423 ± 0,053  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark      s1 * s2 * s5 * s6  thrpt   30  3,384 ± 0,031  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                s3 * s4  thrpt   30  1,741 ± 0,016  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                l2 * s2  thrpt   30  1,125 ± 0,021  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark      l2 * s2 * s5 * s6  thrpt   30  0,513 ± 0,008  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                s1 * l2  thrpt   30  1,119 ± 0,022  ops/ms
BenchmarkDecimalOperators.multiplyBenchmark                l1 * l2  thrpt   30  1,002 ± 0,016  ops/ms
```

I also have the JSON files, of you're interested.